Move formats out of amethyst::assets::formats
* Add amethyst_assets dependency to renderer
* Add submodule to audio module
