+++ src/assets/formats/mod.rs
@@ -1,7 +1,5 @@
 //! Contains formats that amethyst can load as assets.
 
-pub mod audio;
-
 /// Formats to load textures
 pub mod textures;
